==> 03_laboratory_processing.Rmd <==
# Laboratory sample processing

## DNA extraction

### General statistics

```{r dna_extraction, message=F, warning=F}
read_tsv("data/extraction.tsv") %>%
   summarise(
    max= max(extraction_total, na.rm = TRUE),
    min= min(extraction_total, na.rm = TRUE),
    mean= mean(extraction_total, na.rm = TRUE),
    sd = sd(extraction_total, na.rm = TRUE),
    median=median(extraction_total, na.rm = TRUE),
    IQR=IQR(extraction_total, na.rm = TRUE)
  ) %>%
  tt()
```

#### Data distribution

### Sample types

```{r dna_extraction_sample_type_hist, message=F, warning=F, fig.height=6, fig.width=10, fig.fullwidth=TRUE}
left_join(read_tsv("data/extraction.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab"))  %>%
  ggplot(aes(x=sample_type, y= extraction_total, fill=sample_type, color=sample_type)) + 
    ylim(0, 2000) +
    geom_jitter(position = position_jitter(width = 0.2), alpha = 0.5, size=0.5) +
    stat_halfeye(adjust = 0.5,width = 0.5, .width = 0, justification = -.55,normalize = "groups") +
    scale_color_manual(values = c("#bdca50", "#AA3377")) +   
    scale_fill_manual(values = c("#bdca5050", "#AA337750")) +
    labs(y="Density",x="DNA yield", fill="Sample type", color="Sample type") +
    theme_classic()

ggsave("figures/extraction_type.pdf",width=6, height=4, units="in")

left_join(read_tsv("data/extraction.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  group_by(sample_type)  %>%
  summarise(
    mean= mean(extraction_total, na.rm = TRUE),
    sd = sd(extraction_total, na.rm = TRUE),
    median=median(extraction_total, na.rm = TRUE),
    IQR=IQR(extraction_total, na.rm = TRUE)) %>%
  tt()
```

### Taxonomy

```{r dna_extraction_taxonomy_jitter, message=F, warning=F, fig.height=6, fig.width=10, fig.fullwidth=TRUE}
left_join(read_tsv("data/extraction.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab"))  %>%
  mutate(tax_group=factor(tax_group,levels=c("Amphibians","Reptiles","Birds","Bats","Mammals"))) %>% 
  ggplot(aes(y=extraction_total,x=tax_group,color=tax_group,fill=tax_group)) + 
    ylim(0, 2000) +
    geom_jitter(position = position_jitter(width = 0.2), alpha = 0.5, size=0.5) +
    stat_halfeye(adjust = 0.5,width = 0.5, .width = 0, justification = -.55,normalize = "groups") +
    scale_color_manual(values = c("#22883380","#EE667780","#CCBB4480","#66CCEE80","#4477AA80")) +
    scale_fill_manual(values = c("#22883380","#EE667780","#CCBB4480","#66CCEE80","#4477AA80")) +
    labs(y="DNA yield",x="Taxonomic group", color="Sample type") +
    theme_classic()

ggsave("figures/extraction_taxa.pdf",width=9, height=4, units="in")

left_join(read_tsv("data/extraction.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  group_by(tax_group)  %>%
  summarise(
    mean= mean(extraction_total, na.rm = TRUE),
    sd = sd(extraction_total, na.rm = TRUE),
    median=median(extraction_total, na.rm = TRUE),
    IQR=IQR(extraction_total, na.rm = TRUE)) %>%
  tt()
```

#### Comparison

```{r dna_extraction_sample_type_comparison, message=F, warning=F}
left_join(read_tsv("data/extraction.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  dplyr::select(extraction_total,sample_type,tax_group) %>%
  mutate(sample_type = factor(sample_type),tax_group=factor(tax_group)) %>%
  lm(rank(extraction_total) ~ tax_group+sample_type,data=.) %>%
  Anova(.,type="III")%>%
  tidy()%>%
  tt()

left_join(read_tsv("data/extraction.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  dplyr::select(extraction_total,sample_type,tax_group) %>%
  mutate(sample_type = factor(sample_type),tax_group=factor(tax_group)) %>%
  lm(rank(extraction_total) ~ tax_group+sample_type,data=.) %>%
  glht(.,linfct = mcp(tax_group = "Tukey"))%>%
  summary()%>%
  tidy()%>%
  tt()
```

## Sequencing library preparation

### Overall

```{r library_preparation_stats, message=F, warning=F}
left_join(read_tsv("data/library.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  filter(library_PCR_cycles_required > 0) %>%
   summarise(
    max= max(library_input_dna, na.rm = TRUE),
    min= min(library_input_dna, na.rm = TRUE),
    mean= mean(library_input_dna, na.rm = TRUE),
    sd = sd(library_input_dna, na.rm = TRUE),
    median=median(library_input_dna, na.rm = TRUE),
    IQR=IQR(library_input_dna, na.rm = TRUE)) %>%
  tt()

left_join(read_tsv("data/library.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  filter(library_PCR_cycles_required > 0) %>%
   summarise(
    max= max(library_PCR_cycles_required, na.rm = TRUE),
    min= min(library_PCR_cycles_required, na.rm = TRUE),
    mean= mean(library_PCR_cycles_required, na.rm = TRUE),
    sd = sd(library_PCR_cycles_required, na.rm = TRUE),
    median=median(library_PCR_cycles_required, na.rm = TRUE),
    IQR=IQR(library_PCR_cycles_required, na.rm = TRUE)) %>%
  tt()
```

```{r inputdna_vs_pcr_general, message=F, warning=F}
left_join(read_tsv("data/library.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  filter(library_PCR_cycles_required > 0) %>%
  filter(library_input_dna > 0) %>%
  mutate(library_PCR_cycles_required=round(library_PCR_cycles_required))%>%
  glm(library_PCR_cycles_required ~ tax_group*log(library_input_dna)+sample_type*log(library_input_dna), data = .,family=poisson())  %>%
  Anova(.,test.statistic="Wald")

left_join(read_tsv("data/library.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  filter(library_PCR_cycles_required > 0) %>%
  filter(library_input_dna > 0) %>%
  filter(library_input_dna < 200) %>%
  mutate(library_PCR_cycles_required=round(library_PCR_cycles_required))%>%
  glm(library_PCR_cycles_required ~ tax_group*log(library_input_dna)+sample_type*log(library_input_dna), data = .,family=poisson())  %>%
  plot_model(.,type="pred",terms=c("library_input_dna[1:200]","sample_type"),show.data = TRUE,dot.size = 1,line.size = 1)+
  scale_color_manual(values = c("#bdca50", "#AA3377")) +
  scale_fill_manual(values = c("#bdca5050", "#AA337750")) +
  labs(y="Required number of cycles",x="Amount of inputted DNA (ng)", color="Sample type") +
  theme_classic()

ggsave("figures/cycles_type.pdf",width=6, height=3, units="in")

left_join(read_tsv("data/library.tsv"),
          read_tsv("data/sample.tsv"),
          by="sample_id") %>% 
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  filter(library_PCR_cycles_required > 0) %>%
  filter(library_input_dna > 0) %>%
  filter(library_input_dna < 200) %>%
  mutate(library_PCR_cycles_required=round(library_PCR_cycles_required))%>%
  mutate(tax_group=factor(tax_group,levels=c("Amphibians","Reptiles","Birds","Bats","Mammals"))) %>% 
  glm(library_PCR_cycles_required ~ tax_group*log(library_input_dna)+sample_type*log(library_input_dna), data = .,family=poisson())  %>%
  plot_model(.,type="pred",terms=c("library_input_dna [1:200]","tax_group"),show.data = TRUE,dot.size = 1,line.size = 1)+
  scale_color_manual(values = c("#228833","#EE6677","#CCBB44","#66CCEE","#4477AA")) +
        scale_fill_manual(values = c("#22883380","#EE667780","#CCBB4480","#66CCEE80","#4477AA80")) +
        labs(y="Required number of cycles",x="Amount of inputted DNA (ng)", color="Sample type", fill="Sample type") +
  labs(y="Required number of cycles",x="Amount of inputted DNA (ng)", color="Taxonomic group") +
  theme_classic()

ggsave("figures/cycles_taxa.pdf",width=6, height=3, units="in")
```

## Data quality

```{r pcr_vs_lowqual_type, message=F, warning=F, fig.height=6, fig.width=10, fig.fullwidth=TRUE}
left_join(read_tsv("data/preprocessing.tsv"),read_tsv("data/sample.tsv"),by="sample_id") %>%
  left_join(read_tsv("data/sequence.tsv"),by="sequence_id") %>%
  left_join(read_tsv("data/index.tsv"),by="index_id") %>%
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  mutate(lowqual_perc_reads=(1-reads_post_fastp/reads_pre_fastp)*100) %>%
  mutate(lowqual_perc_bases=(1-bases_post_fastp/bases_pre_fastp)*100) %>%
  #Plot map  EE6677 < bats
  ggplot(aes(y=lowqual_perc_bases, x=index_PCR_cycles_given, colour=sample_type, fill=sample_type, group=sample_type)) +
    geom_jitter(alpha=0.3) +
    stat_smooth(method = "gam", formula = y ~ s(x, bs = "ps",k=4),se=FALSE, geom = "smooth", alpha=0.2) +
    scale_color_manual(values = c("#bdca50", "#AA3377")) +   
    scale_fill_manual(values = c("#bdca5080", "#AA337780")) +
    scale_x_log10() +   
    theme_classic() +
    theme(legend.position = "bottom")
```

```{r pcr_vs_lowqual_taxa, message=F, warning=F, fig.height=6, fig.width=10, fig.fullwidth=TRUE}
left_join(read_tsv("data/preprocessing.tsv"),read_tsv("data/sample.tsv"),by="sample_id") %>%
  left_join(read_tsv("data/sequence.tsv"),by="sequence_id") %>%
  left_join(read_tsv("data/index.tsv"),by="index_id") %>%
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  mutate(lowqual_perc_reads=(1-reads_post_fastp/reads_pre_fastp)*100) %>%
  mutate(lowqual_perc_bases=(1-bases_post_fastp/bases_pre_fastp)*100) %>%
  mutate(tax_group=factor(tax_group,levels=c("Amphibians","Reptiles","Birds","Bats","Mammals"))) %>% 
  ggplot(aes(y=lowqual_perc_bases, x=index_PCR_cycles_given, colour=tax_group, fill=tax_group, group=tax_group)) +
    geom_jitter(alpha=0.3) +
    stat_smooth(method = "gam", formula = y ~ s(x, bs = "ps",k=4),se=FALSE, geom = "smooth", alpha=0.2) +
        scale_color_manual(values = c("#228833","#EE6677","#CCBB44","#66CCEE","#4477AA")) +
        scale_fill_manual(values = c("#22883380","#EE667780","#CCBB4480","#66CCEE80","#4477AA80")) +
    scale_x_log10() +   
    theme_classic() +
    theme(legend.position = "bottom")
```

## Data duplication

```{r pcr_vs_duplicates_type_plot, message=F, warning=F, fig.height=6, fig.width=10, fig.fullwidth=TRUE}
left_join(read_tsv("data/preprocessing.tsv"),read_tsv("data/sample.tsv"),by="sample_id") %>%
  left_join(read_tsv("data/sequence.tsv"),by="sequence_id") %>%
  left_join(read_tsv("data/index.tsv"),by="index_id") %>%
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  mutate(lowqual_perc_reads=(1-reads_post_fastp/reads_pre_fastp)*100) %>%
  mutate(lowqual_perc_bases=(1-bases_post_fastp/bases_pre_fastp)*100) %>%
  #Plot map  EE6677 < bats
  ggplot(aes(y=host_duplicate_fraction, x=index_PCR_cycles_given, colour=sample_type, fill=sample_type, group=sample_type)) +
    geom_jitter(alpha=0.3) +
    stat_smooth(method = "gam", formula = y ~ s(x, bs = "ps",k=4),se=FALSE, geom = "smooth", alpha=0.2) +
    scale_color_manual(values = c("#bdca50", "#AA3377")) +   
    scale_fill_manual(values = c("#bdca5080", "#AA337780")) +
    scale_x_log10() +   
    theme_classic() +
    theme(legend.position = "bottom")
```

```{r pcr_vs_duplicates_taxa_plot, message=F, warning=F, fig.height=6, fig.width=10, fig.fullwidth=TRUE}
left_join(read_tsv("data/preprocessing.tsv"),read_tsv("data/sample.tsv"),by="sample_id") %>%
  left_join(read_tsv("data/sequence.tsv"),by="sequence_id") %>%
  left_join(read_tsv("data/index.tsv"),by="index_id") %>%
  filter(sample_type %in% c("Faecal", "Anal/cloacal swab")) %>%
  mutate(lowqual_perc_reads=(1-reads_post_fastp/reads_pre_fastp)*100) %>%
  mutate(lowqual_perc_bases=(1-bases_post_fastp/bases_pre_fastp)*100) %>%
  mutate(tax_group=factor(tax_group,levels=c("Amphibians","Reptiles","Birds","Bats","Mammals"))) %>% 
  ggplot(aes(y=host_duplicate_fraction, x=index_PCR_cycles_given, colour=tax_group, fill=tax_group, group=tax_group)) +
    geom_jitter(alpha=0.3) +
    stat_smooth(method = "gam", formula = y ~ s(x, bs = "ps",k=4),se=FALSE, geom = "smooth", alpha=0.2) +
        scale_color_manual(values = c("#228833","#EE6677","#CCBB44","#66CCEE","#4477AA")) +
        scale_fill_manual(values = c("#22883380","#EE667780","#CCBB4480","#66CCEE80","#4477AA80")) +
    scale_x_log10() +   
    theme_classic() +
    theme(legend.position = "bottom")
```
